Added a META-INF/beans.xml to mark this as a bean archive.
This is not because this library SHOULD be a bean archive,
but Payara in particular has problems with archives that
do not have META-INF/beans.xmls in them (!).  At any rate,
introducing this does no harm, and the value of its
bean-discovery-mode is annotated, and nothing in this library
actually has any bean-defining annotations, so it's effectively
a no-op.